Moved 'lib' and 'constants' folder into LoomSE
