+++ source/LoomSE/constants/applicationActions.js
