+++ source/LoomSE/constants/eventActions.js
