+++ source/LoomSE/constants/eventStates.js
