+++ source/LoomSE/constants/plainTextFiles.js
