+++ source/LoomSE/constants/videoSourceTypes.js
