+++ source/LoomSE/lib/browser/FullScreen.js
