+++ source/LoomSE/lib/browser/LocalStorage.js
