+++ source/LoomSE/lib/browser/index.js
